requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
h5py==3.7.0
hdf5storage==0.1.18
idna==3.3
numpy==1.22.4
opencv-python==4.6.0.66
Pillow==9.1.1
requests==2.28.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
typing-extensions==4.2.0
urllib3==1.26.9
